:sparkles: [FEAT] (#10) FriendUiModel 구현
